refactor. 컴포넌트 단위가 커짐에 따라 Description으로 컴포넌트 분리
[#14]
